requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
catalystwan>=0.33.7
gmqtt>=0.6.16
urllib3
paho-mqtt